Included suggestions from https://blueprints.launchpad.net/hugin/+spec/hugin-appimage
Added exec wrapper

Improved AppRun script

+++ build-appimage.sh
@@ -25,7 +25,8 @@ echo "Building and installing vigra"
 echo ""
 
 cd /work
-rm -rf vigra
+if [ ! -e vigra ] ; then
+#rm -rf vigra
 git clone https://github.com/ukoethe/vigra.git || exit 1
 cd vigra || exit 1
 rm -rf vigra-build
@@ -33,7 +34,7 @@ mkdir -p vigra-build || exit 1
 cd vigra-build || exit 1
 cmake3 -DWITH_OPENEXR=1  -DCMAKE_INSTALL_PREFIX=/usr/local ../ || exit 1
 make -j 3 install || exit 1
-
+fi
 
 echo ""
 echo "########################################################################"
@@ -42,7 +43,8 @@ echo "Building and installing enblend"
 echo ""
 
 cd /work
-rm -rf enblend
+if [ ! -e enblend ]; then
+#rm -rf enblend
 hg clone http://hg.code.sf.net/p/enblend/code enblend || exit 1 
 cd enblend || exit 1
 git clone https://github.com/akrzemi1/Optional.git Optional-master || exit 1
@@ -51,6 +53,7 @@ cd enblend-build || exit 1
 #CXXFLAGS='-std=gnu++1y' CPPFLAGS='-I/opt/rh/devtoolset-4/root/usr/include/c++/5.3.1/experimental' 
 cmake3 -DSOURCE_BASE_DIR="/work/enblend" -DCMAKE_INSTALL_PREFIX=/usr/local ../ || exit 1
 make -j 3 install || exit 1
+fi
 cd ..
 
 
@@ -61,10 +64,12 @@ echo "Building and installing librsvg"
 echo ""
 
 export PATH=$HOME/.cargo/bin:$PATH
+if [ ! -e /work/librsvg-2.40.16 ]; then
 (cd /work && curl https://sh.rustup.rs -sSf > ./r.sh && bash ./r.sh -y && \
 rm -rf librsvg* && wget http://ftp.gnome.org/pub/gnome/sources/librsvg/2.40/librsvg-2.40.16.tar.xz && \
 tar xvf librsvg-2.40.16.tar.xz && cd librsvg-2.40.16 && \
 ./configure --prefix=/usr/local && make -j 3 install) || exit 1
+fi
 
 
 
@@ -75,7 +80,8 @@ echo "Building and installing wxWidgets"
 echo ""
 
 cd /work
-rm -rf wxWidgets-*
+if [ ! -e wxWidgets-3.0.4 ]; then
+#rm -rf wxWidgets-*
 wget https://github.com/wxWidgets/wxWidgets/releases/download/v3.0.4/wxWidgets-3.0.4.tar.bz2 || exit 1
 tar xvf wxWidgets-3.0.4.tar.bz2 || exit 1
 cd wxWidgets-3.0.4 || exit 1
@@ -83,7 +89,7 @@ mkdir -p wx-build || exit 1
 cd wx-build || exit 1
 ../configure --enable-unicode --with-opengl --prefix=/usr/local || exit 1
 make -j 3 install || exit 1
-
+fi
 
 echo ""
 echo "########################################################################"
@@ -91,8 +97,9 @@ echo ""
 echo "Building and installing hugin"
 echo ""
 
+if [ ! -e /work/hugin ]; then
 cd /work
-rm -rf hugin
+#rm -rf hugin
 hg clone http://hg.code.sf.net/p/hugin/hugin hugin || exit 1
 cd hugin || exit 1
 patch -N -p1 < /sources/hugin-bundle.patch || exit 1
@@ -102,9 +109,11 @@ cd hugin-build || exit 1
 cmake3 -DCMAKE_INSTALL_PREFIX=/usr/local \
 	-DCMAKE_BUILD_TYPE=Release \
 	-DBUILD_HSI:BOOL=OFF \
+	-DUSE_GDKBACKEND_X11:BOOL=ON \
 	-DUNIX_SELF_CONTAINED_BUNDLE:BOOL=ON ../ || exit 1
 make -j 3 install || exit 1
 cd ..
+fi
